+++ apps/api-java/src/main/java/org/opencivic/signalos/domain/InterfaceMode.java
@@ -1,0 +1,6 @@
+package org.opencivic.signalos.domain;
+
+public enum InterfaceMode {
+    SIMPLE,
+    ADVANCED
+}